Fix 1 loss in losses handover due to fragmentation in txed/retx buffers.
Since txed/retx buffers might contain PDUs that are a fragment
of the original RLC SDU. Particularly, when forwarding txonBuffer
to target eNB, the front of the txonBuffer could be a segment
that is fragmented (a part of it is already become PDUs). This
incomplete segment has 0 sequence number and will be discard
during forwarding (can't be recognized as packet SDU).
The txed/retx buffers also contains PDUs that are not able to
be reconstructed to get the original SDU. For example, UE
may acked some PDUs of a SDU so txed/retx buffer only
store PDUs that are together form only a portion of the SDU.

To solve this issue:
(1) A complete SDU of the last segemented SDU in txonBuffer is
stored and this complete SDU will be forwarded to target eNB.
(2) Second issue when txed/retx buffer do not have enough PDUs
to construct a SDU (some PDUs are acked) will be solved in next commit.