HO for RlcUm and RlcAm (downlink) complete.
1. RlcUm HO: unlike the seamless HO implemented in the
simulator, RlcUm seamless HO as described in text book
and specs (10.1.2.1 TS 136 300) should forward buffered
PDCP PDUs to the target eNB. The seamless HO in the
simulator discards the RlcTxonBuffer which is essentially
the PDCP TxPDUs. This implementation follows the specs
in which it forwards RlcTxonBuffer (stores not transmitted
but bufferred PDCP PDUs) to the target eNB.

2. RlcAm HO: new in the simulator. This follows the
specs (10.1.2.1 TS 136 300) and done for DOWNLINK.
This lossless HO forward both bufferred PDCP PDUs and
transmitted but not yet acknowledged PDCP PDUs.
Limitations: "end marker" from SGW that marks
the end of packet stream that destined for the
source eNB is not implemented. This makes the
in-order delivery at the target eNB not guarantee
(Although from auto HO experiments, there is no
evidence showing this out-of-order at the target
eNB issue).